avoid env setup in notification id tests
